+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Fix `Enumerable` implementation crashing with `CaseClauseError` on read errors.
   * Fix `Enumerable.count/1` returning invalid protocol value on error.
   * Fix `Enumerable.slice/1` crashing with `MatchError` when object is invalid.
+  * Fix `Enumerable` slicing function ignoring seek/read errors.
 
 ## v0.2.6 - 2026-05-01